refactor translators
