+++ dsf-pseudonymization/dsf-pseudonymization-medic/src/main/java/org/highmed/pseudonymization/translation/ResultSetTranslatorToTtpCreateRbfOnlyDropOtherColumns.java
@@ -1,0 +1,5 @@
+package org.highmed.pseudonymization.translation;
+
+public interface ResultSetTranslatorToTtpCreateRbfOnlyDropOtherColumns extends ResultSetTranslatorToTtp
+{
+}
