+++ dsf-pseudonymization/dsf-pseudonymization-medic/src/main/java/org/highmed/pseudonymization/translation/ResultSetTranslatorToTtpEncrypt.java
@@ -1,0 +1,5 @@
+package org.highmed.pseudonymization.translation;
+
+public interface ResultSetTranslatorToTtpEncrypt extends ResultSetTranslatorToTtp
+{
+}
