--- dsf-pseudonymization/dsf-pseudonymization-medic/src/main/java/org/highmed/pseudonymization/translation/ResultSetTranslatorToTtpNoRbf.java
@@ -1,5 +1,0 @@
-package org.highmed.pseudonymization.translation;
-
-public interface ResultSetTranslatorToTtpNoRbf extends ResultSetTranslatorToTtp
-{
-}
