--- dsf-pseudonymization/dsf-pseudonymization-medic/src/main/java/org/highmed/pseudonymization/translation/ResultSetTranslatorToTtpRbfOnly.java
@@ -1,5 +1,0 @@
-package org.highmed.pseudonymization.translation;
-
-public interface ResultSetTranslatorToTtpRbfOnly extends ResultSetTranslatorToTtp
-{
-}
