--- dsf-pseudonymization/dsf-pseudonymization-medic/src/main/java/org/highmed/pseudonymization/translation/ResultSetTranslatorToTtpWithRbf.java
@@ -1,5 +1,0 @@
-package org.highmed.pseudonymization.translation;
-
-public interface ResultSetTranslatorToTtpWithRbf extends ResultSetTranslatorToTtp
-{
-}
