+++ dsf-pseudonymization/dsf-pseudonymization-medic/src/test/resources/result_6.json
@@ -1,0 +1,46 @@
+{
+   "meta":{
+      "_href":"http://test.com",
+      "_type":"RESULTSET",
+      "_schema_version":"1.0.0",
+      "_created":"2017-08-19T00:25:47.568+02:00",
+      "_generator":"DIPS.OpenEhr.ResultSets.Serialization.Json.ResultSetJsonWriter (5.0.0.0)",
+      "_executed_aql":"SELECT e FROM EHR(e)"
+   },
+   "name":"<the name/identifier of the stored query that was executed>",
+   "q":"select e/ehr_status/subject/external_ref/id/value, c/context/start_time/value as startTime,  c/uid/value as cid, c/name from EHR e contains  Composition c limit 2",
+   "columns":[
+      {
+         "name":"EHRID",
+         "path":"/ehr_status/subject/external_ref/id/value"
+      },
+      {
+         "name":"RBF",
+         "path":"/rbf/value"
+      },
+      {
+         "name":"startTime",
+         "path":"/context/start_time/value"
+      },
+      {
+         "name":"cid",
+         "path":"/uid/value"
+      },
+      {
+         "name":"#3",
+         "path":"/name"
+      }
+   ],
+   "rows":[
+      [
+         "ehrId1",
+         "EM4AAQBiIWJHrYA0BClKbCCWESYFygQokRiqhEKmjYCEjMCAAiEwlEakZgMTFY08AGCpKwCcAikXLMZLFYEmhAKABKySABnI1iAgJlzEICHhiaUIgGIix0dBk+gIYxAoBIEFBbpIAWPCAO4lIGIcStCDg6yQMZApJAko60gSSRCxbcimkSkCwCHPQnoQqYAIUpSexA4WIbA0cpB1FqTAx4A8I3fACjFWpShTSTO8gIsqU4ClrBSFsmxUsMJoQsBAE6pNDQ4kHCsRHAJLAc4BF9uTBFqAA4cIAEidKCAMoozY",
+         "2017-02-16T13:50:11.308+01:00",
+         "90910cf0-66a0-4382-b1f8-c0f27e81b42d::default::1",
+         {
+            "_type":"DV_TEXT",
+            "value":"Labs"
+         }
+      ]
+   ]
+}
